requirements.txt
numpy
pandas
tqdm
typing-extensions 
google-generativeai
Pillow  
ratelimit
PyMuPDF 
python-dotenv 
streamlit